Добавил проверку на ошибки преобразований

+++ src/prlib/control.h
@@ -5,7 +5,7 @@
 void bound(unsigned *t, unsigned *f, unsigned *j, TClVar *ClVar, array *heap);
 ErrorCode control(TScVar *ScVar, TClVar *ClVar, array *heap, bool *EnableRunning);
 bool unify(unsigned tt1, unsigned tt2, unsigned ff1, unsigned ff2, TClVar *, array *);
-unsigned char calculation(unsigned term, unsigned frame, long *i, float *f, TClVar *ClVar, array *heap);
+unsigned char calculation(unsigned term, unsigned frame, IntegerType *i, FloatType *f, TClVar *ClVar, array *heap);
 void outerror(ErrorCode err);
 void zero(TClVar *ClVar);
 void to_stac(unsigned *b, unsigned index, unsigned value);
